+++ components/Actuators/HighLevel/loaderLogic.py
@@ -1,8 +1,8 @@
 from robotMap import XboxMap
-from components.intakeMotor import IntakeMotor
-from components.hopperMotor import HopperMotor
+from components.Actuators.LowLevel.intakeMotor import IntakeMotor
+from components.Actuators.LowLevel.hopperMotor import HopperMotor
 from components.breakSensors import Sensors, State
-from components.feederMap import FeederMap, Type
+from components.Actuators.HighLevel.feederMap import FeederMap, Type
 from utils.DirectionEnums import Direction
 from magicbot import StateMachine, state, timed_state, tunable, feedback
